% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AVE_hand.R
\name{AVE}
\alias{AVE}
\title{AVE by hnad}
\usage{
AVE(x, digit = 3, markdown = FALSE)
}
\arguments{
\item{x}{lavaan result}

\item{digit}{rounding}

\item{markdown}{data.frame or markdown. default FALSE}
}
\description{
AVE by hnad
}
