% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_lme_model.R
\name{bind_lme_model}
\alias{bind_lme_model}
\title{Tbind_lme_model}
\usage{
bind_lme_model(
  ...,
  model_name = paste("model_", 1:length(model)),
  title = "Comparison of model"
)
}
\arguments{
\item{...}{model name list}

\item{model_name}{auto name, model}

\item{title}{title input}
}
\value{
table data of comparison model
}
\description{
Tbind_lme_model
}
\examples{
\dontrun{
bind_lme_model(gmlm.pois.m1,  gmlm.pois.m2a,  gmlm.pois.m2b)
}
}
