% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_summary_other.R
\name{bind_pois}
\alias{bind_pois}
\title{The function combining the Poisson analysis result}
\usage{
bind_pois(model1, model2)
}
\arguments{
\item{model1}{model 1 result}

\item{model2}{model 2 ruslt}
}
\value{
compare table
}
\description{
The function combining the Poisson analysis result
}
