% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa2.R
\name{cfa2}
\alias{cfa2}
\title{Create CFA result summary by lavaan sem estimated result}
\usage{
cfa2(
  x,
  format = "markdown",
  dataset = NA,
  model = NA,
  cut = 0.7,
  angle = 90,
  cex = 11,
  hjust = 0.9,
  val.size = 4,
  dis.sort = TRUE,
  rename = F,
  var_name = NA,
  digits = 3,
  res = "all"
)
}
\arguments{
\item{x}{lavaan result}

\item{format}{knitr markdown decision}

\item{dataset}{this is import name that dataset column name for HTMT}

\item{model}{lavaan model need indices validation name for HTMT}

\item{cut}{loading criteran}

\item{angle}{tesxt angle}

\item{cex}{text size}

\item{hjust}{on text value gap}

\item{val.size}{text size}

\item{dis.sort}{item sort}

\item{rename}{item validiation new name apply}

\item{var_name}{new text name default NULL}

\item{digits}{value rounding}

\item{res}{result type}
}
\description{
Create CFA result summary by lavaan sem estimated result
}
