% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Char2num.R
\name{Char2num}
\alias{Char2num}
\title{character datq tranfomation double}
\usage{
Char2num(data, colnumber = "auto", iter = NULL)
}
\arguments{
\item{data}{data.frame or matrix}

\item{colnumber}{'auto', 'manual'}

\item{iter}{iteration n}
}
\value{
char to numeric
}
\description{
character datq tranfomation double
}
