% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compare_Fit.R
\name{Compare_Fit}
\alias{Compare_Fit}
\title{model fit}
\usage{
Compare_Fit(..., option = 2, format = "markdown")
}
\arguments{
\item{...}{lavaan result}

\item{option}{option =1 markdonw, 2 data.frame}

\item{format}{"markdown', 'html'}
}
\value{
paper format
}
\description{
model fit
}
