% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compare_grp_effect.R
\name{Compare_grp_effect}
\alias{Compare_grp_effect}
\title{Comparative analysis of direct effects by group}
\usage{
Compare_grp_effect(
  global,
  groupsem,
  digit = 3,
  title = "",
  g1 = "g1",
  g2 = "g2",
  namestart = 8
)
}
\arguments{
\item{global}{lavaan overall group analysis results}

\item{groupsem}{Analysis results by lavaan group}

\item{digit}{3 round}

\item{title}{title}

\item{g1}{group name 1}

\item{g2}{group name 2}

\item{namestart}{group name extraction}
}
\value{
compare table
}
\description{
Comparative analysis of direct effects by group
}
