% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_path.R
\name{create_path}
\alias{create_path}
\title{Function to create path plspm}
\usage{
create_path(
  row_names = c("r1", "r2", "r3"),
  col_names = NULL,
  relationship = NULL
)
}
\arguments{
\item{row_names}{row names c("r1","r2","r3")}

\item{col_names}{col names}

\item{relationship}{form var1, to var2}
}
\value{
matrix
}
\description{
Function to create path plspm
}
\examples{

\dontrun{
edu_path <- create_path(
  row_names = c("Support", "Advising", "Tutoring", "Value", "Satisfaction", "Loyalty"),
  col_names = c("Support", "Advising", "Tutoring", "Value", "Satisfaction", "Loyalty"),
  relationship = list(
    list(from = "Value",
         to = c("Support", "Advising", "Tutoring")),
    list(from = "Satisfaction",
         to = c("Support", "Advising", "Tutoring","Value")),
    list(from = "Loyalty",
         to =c("Satisfaction"))
  )
)
edu_path


edu_path <- create_path(
  row_names = c("Support", "Advising", "Tutoring", "Value", "Satisfaction", "Loyalty"),
  col_names = c("Support", "Advising", "Tutoring", "Value", "Satisfaction", "Loyalty"),
  relationship = list(
    path(from = "Value",
         to = c("Support", "Advising", "Tutoring")),
    path(from = "Satisfaction",
         to = c("Support", "Advising", "Tutoring","Value")),
    path (from = "Loyalty",
         to =c("Satisfaction"))
  )
)
edu_path

}



}
