% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des_stat.R
\name{des_stat}
\alias{des_stat}
\title{Tmean and sd}
\usage{
des_stat(x, digit = 3)
}
\arguments{
\item{x}{data.frame}

\item{digit}{default 3}
}
\value{
reuren mean(sd) type
}
\description{
Tmean and sd
}
\examples{
\dontrun{
df = data.frame(
subject = c(1:6),
mon1 = c(45, 42, 36, 39, 51, 44),
mon3 = c(50, 42, 41, 35, 55, 49),
mon6 = c(55, 45, 43, 40, 59, 56))
df[,2:4] \%>\% des_stat()
}
}
