% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Freq_table.R
\name{Freq_table}
\alias{Freq_table}
\title{Freq_table  Frequency output function------------}
\usage{
Freq_table(data)
}
\arguments{
\item{data}{data.frame}
}
\description{
Freq_table  Frequency output function------------
}
