% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICC_logit.R
\name{ICC.logit}
\alias{ICC.logit}
\title{ICC.logit generalized linear model ICC}
\usage{
ICC.logit(gmlm.dichotomous)
}
\arguments{
\item{gmlm.dichotomous}{logit model analysis result}
}
\value{
ICCresult
}
\description{
ICC.logit generalized linear model ICC
}
\examples{
\dontrun{
lme4::glmer(am ~ wt + (wt |vs ), mtcars,
            family=binomial(link='logit'))|>
            ICC.logit()
}
}
