% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICC.R
\name{ICC}
\alias{ICC}
\title{ICC}
\usage{
ICC(lmedata)
}
\arguments{
\item{lmedata}{lmer result}
}
\description{
ICC
}
\examples{
\dontrun{
data(sleepstudy)
sleepstudy |> str()
lmer(Reaction ~ Days + (Days | Subject), sleepstudy) |> ICC()


}
}
