% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICC_logit.R
\name{ICC_dich}
\alias{ICC_dich}
\title{ICC_dich is generalized linear model ICC}
\usage{
ICC_dich(gmlm.dichotomous)
}
\arguments{
\item{gmlm.dichotomous}{logit model analysis result}
}
\value{
ICCresult
}
\description{
ICC_dich is generalized linear model ICC
}
\examples{
\dontrun{
lme4::glmer(am ~ wt + (wt |vs ), mtcars,
            family=binomial(link='logit'))|>
            ICC_logit()
}
}
