% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LavaanPlot.R
\name{LavaanPlot}
\alias{LavaanPlot}
\title{Lavaan plot draw}
\usage{
LavaanPlot(
  x,
  rotation = "LR",
  coefs = TRUE,
  col = "gray60",
  stand = FALSE,
  label = NULL,
  covs = FALSE,
  stars = c("covs", "latent", "regress"),
  overlap = FALSE,
  layout = "dot",
  digits = 3
)
}
\arguments{
\item{x}{lavaan result}

\item{rotation}{rotation 'LR','RL', 'TB','BT'}

\item{coefs}{TRUE}

\item{col}{default 'gray60'}

\item{stand}{std coeff}

\item{label}{input name labels=c(f1_1="exp", f1_2="want",f1_3="problem") etc}

\item{covs}{TRUE}

\item{overlap}{default FALSE}

\item{layout}{'dot', 'neato', 'fdp', 'sfdp', 'circo', 'twopi', 'nop','nop2', 'osage', 'patchwork'}

\item{digits}{digits = 3#'}
}
\value{
plot
}
\description{
Lavaan plot draw
}
