% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_report.R
\name{lme_report}
\alias{lme_report}
\title{linear mixed model report}
\usage{
lme_report(lmedata, apa = FALSE, fit_more = FALSE)
}
\arguments{
\item{lmedata}{lmedata is lmer() function result}

\item{apa}{= default FALSE, if you  REML is FALSE}

\item{fit_more}{= default FALSE, TRUE detailed report}
}
\description{
linear mixed model report
}
\examples{
\dontrun{
data(sleepstudy)
sleepstudy |> str()
lmer(Reaction ~ Days + (Days | Subject), sleepstudy) |> lme_report()
lmer(Reaction ~ Days + (Days | Subject), sleepstudy) |> lme_report(apa=TRUE)


book.url <- "https://stat.ethz.ch/~meier/teaching/book-anova"
quality <- readRDS(url(file.path(book.url, "data/quality.rds")))
str(quality)
fit.quality <- lmer(score ~ (1 | employee) + (1 | batch) +
                     (1 | employee:batch), data = quality)
summary(fit.quality)
fit.quality|> lme_report()
fit.quality|> lme_report(apa= TRUE) #error the app is FALSE
fit.quality|> lme_report(apa=FALSE) #error the app is FALSE
fit.quality|> ICC()

}

}
