% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdwon_table.R
\name{markdown_table}
\alias{markdown_table}
\title{A tool that outputs data to the viewer}
\usage{
markdown_table(
  lm_data,
  caption = "<Table  > capiton title",
  digits = 2,
  font_size = 18,
  full_width = F,
  table = "classic",
  show = "data",
  format = "html",
  row.names = NA,
  col.names = NA,
  align = NULL,
  label = NULL,
  format.args = list(),
  escape = TRUE,
  table.attr = "",
  booktabs = FALSE,
  longtable = FALSE,
  valign = "t",
  position = "",
  centering = TRUE
)
}
\arguments{
\item{lm_data}{data.frame or etc data}

\item{caption}{table title}

\item{digits}{table values roundings}

\item{font_size}{table font size}

\item{full_width}{table wide and fit values}

\item{table}{tbble type 'paper', 'classic', 'dark', 'minimal'}

\item{show}{'markdown' is output to the viewer, and 'data' is output to the console.}

\item{format}{'markdown' and 'html'. default 'html'.}

\item{row.names}{row names in table(Set if necessary)}

\item{col.names}{column names in table (Set if necessary)}

\item{align}{Table value align}

\item{label}{Table label}

\item{format.args}{format.args}

\item{escape}{escape}

\item{table.attr}{table.attr}

\item{booktabs}{booktabs}

\item{longtable}{longtable}

\item{valign}{valign}

\item{position}{position}

\item{centering}{Align cener in Table}
}
\description{
A tool that outputs data to the viewer
}
\examples{
# example code

  \dontrun{
  markdown_table(mtcars[1:5,], table="minimal")
  }
}
