% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdwon_table.R
\name{markdown_table_s}
\alias{markdown_table_s}
\title{A tool that outputs data to the viewer}
\usage{
markdown_table_s(
  data,
  caption = "html to markdown",
  full_width = FALSE,
  font_size = 20,
  row.names = NA,
  col.names = NA,
  centering = TRUE,
  digits = 3,
  show = "markdown"
)
}
\arguments{
\item{data}{data.frame or etc data}

\item{caption}{table title}

\item{full_width}{table wide and fit values}

\item{font_size}{table font size}

\item{row.names}{row names in table(Set if necessary)}

\item{col.names}{column names in table (Set if necessary)}

\item{centering}{Align cener in Table}

\item{digits}{table values roundings}

\item{show}{'markdown' is output to the viewer, and 'kbl' is output to the console.}
}
\description{
A tool that outputs data to the viewer
}
