% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds_data.R
\name{mds_data}
\alias{mds_data}
\title{MDS data output}
\usage{
mds_data(
  mds_data,
  show = "all",
  variable = c("ZOOM", "Webex", "googleMEET", "Youtube", "NaverBand", "Prism"),
  addxy = 0.5,
  linetype = 1,
  linecolor = "gray40",
  linewidth = 0.8,
  xl = -3,
  xh = 3,
  yl = -2.3,
  yh = 2,
  text_size = 5,
  size.x = 14,
  caption = "MDS"
)
}
\arguments{
\item{mds_data}{mda_data input}

\item{show}{'all','cmdscale','isoMDS','cmdplot_data','isoplot_data', 'plot_data', 'cmd_plot','iso_plot', 'shepard_plot', 'image_plot', 'iso_plot_line', 'iso_line', 'iso_line_auto', 'cmd_plot_line', 'cmd_line', 'cmd_line_auto', 'cmd_gof', 'stress','shepard_data',  'image_data', 'shepardimage_data', 'mat', 'data_mat','scale', 'scale_mat', 'dist','dist_mat','des', 'gof', 'index'}

\item{variable}{ex c("ZOOM","Webex","googleMEET", "Youtube","NaverBand","Prism")}

\item{addxy}{default 0.5}

\item{linetype}{linetype 1}

\item{linecolor}{linecolor}

\item{linewidth}{linewidth= 0.8}

\item{xl}{xl = -3}

\item{xh}{xh = 3}

\item{yl}{yl = -2.3}

\item{yh}{yh = 2}

\item{text_size}{text_size = 5}

\item{size.x}{text_size = 5}

\item{caption}{caption="MDS"}
}
\value{
MDS result overall
}
\description{
MDS data output
}
\examples{

\dontrun{
# mds_onl_i \%>\% mds_data()
mds_onl_i \%>\% mds_data(show="all")

mds_onl_i \%>\% mds_data(show="data_mat")
mds_onl_i \%>\% mds_data(show="data_mat") \%>\% dist()
mds_onl_i \%>\% mds_data(show="dist_mat")
mds_onl_i \%>\% mds_data(show="dist_mat") \%>\% as.matrix() \%>\%
  markdown_table("Distance matrix", digits=3)
}
}
