% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdsmakePlot_.R
\name{mds_makePlot}
\alias{mds_makePlot}
\title{New MDS function  plot comparison ----}
\usage{
mds_makePlot(
  mds_data,
  grid_title = "",
  xl = -3,
  xh = 3,
  yl = -2.3,
  yh = 2,
  variable = c("ZOOM", "Webex", "googleMEET", "Youtube", "naveBand", "prism"),
  show = "manual",
  linecolor = "gray50",
  linetype = 1,
  linewidth = 1,
  what = ""
)
}
\arguments{
\item{mds_data}{mds analysis result data}

\item{grid_title}{your title}

\item{xl}{graph position x low default -3}

\item{xh}{graph position x high default 3}

\item{yl}{graph position y low default -2.3}

\item{yh}{graph position y high default 2}

\item{variable}{current 6. c("ZOOM","Webex","googleMEET", "Youtube","naveBand","prism")}

\item{show}{"manual","auto","line"}

\item{linecolor}{line gray50}

\item{linetype}{type 1}

\item{linewidth}{widtih 1}

\item{what}{input your mind}
}
\description{
New MDS function  plot comparison ----
}
\examples{

ma_new(8- mds_onl[1:20,], grid_title = "2020 full group awareness ")
# mds_onl_2020
mds_onl_i
mds_onl_i \%>\% slice(1:64) \%>\%
  mds_makePlot(grid_title = gtl("total percetption "))




}
