% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanCentering_df.R
\name{meanCentering_grp}
\alias{meanCentering_grp}
\title{meancentering mixed model group}
\usage{
meanCentering_grp(data, grp_id, gv, ...)
}
\arguments{
\item{data}{data.frame}

\item{grp_id}{group id}

\item{gv}{group variable, Group Variables That Separate Groups}

\item{...}{individual measurement independent variables evaluated at individual level}
}
\description{
meancentering mixed model group
}
\examples{
\dontrun{
meanCentering_df(data = mtcars[,c("cyl","mpg","wt","hp")],
                 gv="cyl","mpg","wt","hp")|> data.frame()
}



}
