% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasurementInvariance.R
\name{MeasurementInvariance}
\alias{MeasurementInvariance}
\title{Title}
\usage{
MeasurementInvariance(model, data, grp = "")
}
\arguments{
\item{model}{lavaan mode syntax}

\item{data}{data.frame}

\item{grp}{setting compare group name}
}
\value{
Measurement invariance Testing
}
\description{
Title
}
