% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mga_comparison_plot.R
\name{mga_comparison_plot}
\alias{mga_comparison_plot}
\title{Group comparison and graph}
\usage{
mga_comparison_plot(data, grp1 = "grp1_est", grp2 = "grp2_est", res = "all")
}
\arguments{
\item{data}{seminr group result :  estimate_pls_mga}

\item{grp1}{first group name}

\item{grp2}{second group name}

\item{res}{'all'.'plot','plot_data','res_data'}
}
\value{
graph and data
}
\description{
Group comparison and graph
}
\examples{
\dontrun{
library(seminr)
data(stat_onl)

### model
library(seminr)
lmodel1 <- constructs(
  composite("SRL_meta", single_item("S_Review"), weights = mode_A),
  composite("SRL_meta", single_item("S_Add_learn"), weights = mode_A),
  composite("SRL_meta", single_item("S_Feedback"), weights = mode_A),
  composite("SRL_meta", single_item("S_Focus_on"), weights = mode_A),

  composite("SRL_env", single_item("SE_Place"), weights = mode_A),
  composite("SRL_env", single_item("SE_Time"), weights = mode_A),

 composite("OnlineSAT", single_item("On_Joy"), weights = mode_A),
  composite("OnlineSAT", single_item("On_Easy"), weights = mode_A),
  composite("OnlineSAT", single_item("On_Satisfy"), weights = mode_A),
  composite("OnlineSAT", single_item("On_Engage" ), weights = mode_A),

  composite("USEIntent", single_item("Intension_use"), weights = mode_A),
  composite("AppSatisfy", single_item("upgrade"), weights = mode_A),
  composite("AppSatisfy", single_item("satisfy"), weights = mode_A)
)
## structure model ---------------
lstr1 = relationships(
  paths(from=c("AppSatisfy"), to=c("SRL_meta", "SRL_env" ,"OnlineSAT")),
  paths(from=c("SRL_meta", "SRL_env"),to=c( "OnlineSAT")),
  paths(from=c("SRL_meta", "SRL_env"),to=c( "USEIntent")),
  paths(from=c("OnlineSAT"),to=c( "USEIntent")))

## model estimate ----------------
srlapp_pls  <- estimate_pls(data = stat_onl,
                            measurement_model = lmodel1,
                            structural_model = lstr1)

##group comparison
srlapp_pls_gender <-  estimate_pls_mga(srlapp_pls,
                         stat_onl$gender == "male", cores=6 )
###  Group Comparison between genders
srlapp_pls_gender \%>\% mga_comparison_plot("female","male")
srlapp_pls_gender \%>\%
  mga_comparison_plot("female","male","res_data") \%>\%
  markdown_table("Comparison of path coefficients between gender groups",
                 digits = 3 )


}


}
