% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydescriptive.R
\name{mydes}
\alias{mydes}
\title{mydes 'n','mean','sd','min','max'}
\usage{
mydes(myvariable, digits = 2)
}
\arguments{
\item{myvariable}{variable vector}

\item{digits}{digits}
}
\description{
mydes 'n','mean','sd','min','max'
}
\examples{
\dontrun{

mydes(mtcars$mpg)

}

}
