% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summarise.R
\name{mySummarise}
\alias{mySummarise}
\title{descriptive statistics}
\usage{
mySummarise(x, type = "normalitytest", digit = 2)
}
\arguments{
\item{x}{data.frame or etc}

\item{type}{result option. default 'normalitytest'. if you possible select optuon. 'col', 'row','skew','kurt','mean','sd', 'n', 'median', 'all', 'normalitytest','normalitytest_data', 'df', 'mvn','shapiro', 'ref'}

\item{digit}{rounding default 2}
}
\description{
descriptive statistics
}
\examples{
#example code
\dontrun{
 mtcars \%>\% mySummarise("col", 2)
 mtcars \%>\% mySummarise("mean",4)
 mtcars \%>\% mySummarise("sd",3)
 mtcars \%>\% mySummarise("skew",4)
 mtcars \%>\% mySummarise("kurt",4)
 mtcars \%>\% mySummarise("all")
 mtcars \%>\% mySummarise("all_tibble")
 mtcars \%>\% mySummarise("all", 3)
 mtcars \%>\% mySummarise("all", 4)
 mtcars \%>\% mySummarise("normalitytest")
 mtcars \%>\% mySummarise("normalitytest", 3)
 mtcars \%>\% mySummarise("normalitytest_data")

}


}
