% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydescriptive.R
\name{mysummary}
\alias{mysummary}
\title{my summary descriptive statistics}
\usage{
mysummary(myobject, myvariable)
}
\arguments{
\item{myobject}{data.frame, matrix}

\item{myvariable}{column variable}
}
\value{
size   MEAN    SD  MIN  MAX
}
\description{
my summary descriptive statistics
}
\examples{
\dontrun{
mysummary(mtcars, "mpg")
}
}
