% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysummaryBy.R
\name{mysummaryBy}
\alias{mysummaryBy}
\title{statistics by group}
\usage{
mysummaryBy(formula, data, add_var = NULL)
}
\arguments{
\item{formula}{general formula}

\item{data}{data.frame}

\item{add_var}{one sample name, Used to obtain statistics for one variable}
}
\value{
Mea, SD, N, min, max
}
\description{
statistics by group
}
\examples{
\donrun{
## Used to obtain statistics for one variable
mysummaryBy(mpg ~ 1, data = mtcars)
mysummaryBy(mpg ~ 1, data = mtcars, "mpg")
mysummaryBy(mpg ~ 1, data = mtcars, add_var = "mpg")

##The group variable is 2 or greater
mysummaryBy(mpg ~ vs, data = mtcars)
mysummaryBy(mpg ~ vs+am, data = mtcars)
mysummaryBy(mpg ~ vs+am+cyl, data = mtcars)

}
}
