% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_boot_summary.R
\name{pls_boot_summary}
\alias{pls_boot_summary}
\title{Summary of PLS-SEM bootstrap results}
\usage{
pls_boot_summary(
  pls_boot,
  res = "all",
  title = "",
  boot = FALSE,
  nboot = 100,
  diagram = FALSE,
  rename = FALSE,
  var_name = NULL,
  vjust = -0.5
)
}
\arguments{
\item{pls_boot}{bootstrap data}

\item{res}{'all','loadings','loadings0','loading_bar','paths','plot'.'htmt','browse_plot','varName'}

\item{title}{your tite}

\item{boot}{default FALSE}

\item{nboot}{default 100}

\item{diagram}{default FALSE}

\item{rename}{default FALSE}

\item{var_name}{default NULL}

\item{vjust}{default -0.5}
}
\value{
boot sig result
}
\description{
Summary of PLS-SEM bootstrap results
}
\examples{
\dontrun{
boot_srlapp_pls <- bootstrap_model(srlapp_pls, nboot = 10000, cores = 4)
boot_srlapp_pls
boot_srlapp_pls \%>\%  pls_boot_summary()

boot_srlapp_pls \%>\%  pls_boot_summary(rename = T, var_name = varName_boot)

varName_boot = c(
             "S_Review","S_Add_learn","S_Feedback","S-Focus_on",
             "SE_place","SE_time",
             "On_joy","On_easy","On_satisfy","On_paticipate" ,
             "IntensionUse",
             "A_upgrade","A_satisfy")

}


}
