% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_cfa.R
\name{pls_cfa}
\alias{pls_cfa}
\title{seminr pls-SEM CFA}
\usage{
pls_cfa(
  pls_data,
  res = "all",
  diagram = FALSE,
  reliability_plot = FALSE,
  digits = 3,
  boot = FALSE,
  nboot = 100,
  rename = FALSE,
  var_name = NULL,
  vjust = -0.5
)
}
\arguments{
\item{pls_data}{seminr result}

\item{res}{'all','loadings','loadings_rename','CR_AVE','CR_AVE_sig','FL_criteria','HTMT','crossloadings','f2','reliability_plot','loading_bar','browse_Viewer','effect','mediation','indirect_effect','Indirect_effect','Total_effect','total_effect','Direct_effect','direct_effect','r2','str_cor','paths','loadings_boot','paths_boot','htmt_boot','vif','vifr','vifc','summarise'}

\item{diagram}{deafult FALSE}

\item{reliability_plot}{deafult FALSE}

\item{digits}{deafult 3}

\item{boot}{deafult FALSE}

\item{nboot}{deafult 100}

\item{rename}{deafult FALSE}

\item{var_name}{deafult NULL}

\item{vjust}{deafult -0.5}
}
\value{
variout result
}
\description{
seminr pls-SEM CFA
}
\examples{
\dontrun{
# View result
srlapp_pls \%>\% pls_cfa(rename = F, var_name = varName)
srlapp_pls \%>\% pls_cfa(rename = T, var_name = varName)
srl_boot3 \%>\% pls_boot_summary()
varName = c( "A_upgrade","A_satisfy",
             "S_Review","S_Add_learn","S_Feedback","S-Focus_on",
             "SE_place","SE_time",
             "On_joy","On_easy","On_satisfy","On_paticipate" ,
             "IntensionUse")
}

}
