% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summarise.R
\name{Summarise}
\alias{Summarise}
\title{descriptive statistics}
\usage{
Summarise(x, type = "normalitytest", digit = 2)
}
\arguments{
\item{x}{data.frame or etc}

\item{type}{result option. default 'normalitytest'. if you possible select optuon. 'col', 'row','skew','kurt','mean','sd', 'n', 'median', 'all', 'normalitytest','normalitytest_data', 'df', 'mvn','shapiro', 'ref'}

\item{digit}{rounding default 2}
}
\description{
descriptive statistics
}
\examples{
#example code
\dontrun{
 mtcars\%>\% Summarise("col", 2)
 mtcars \%>\% Summarise("mean",4)
 mtcars \%>\% Summarise("sd",3)
 mtcars \%>\% Summarise("skew",4)
 mtcars \%>\% Summarise("kurt",4)
 mtcars \%>\% Summarise("all")
 mtcars \%>\% Summarise("all_tibble")
 mtcars \%>\% Summarise("all", 3)
 mtcars \%>\% Summarise("all", 4)
 mtcars \%>\% Summarise("normalitytest")
 mtcars \%>\% Summarise("normalitytest", 3)
 mtcars \%>\% Summarise("normalitytest_data")

}


}
