% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_summary_other.R
\name{summary_pois}
\alias{summary_pois}
\title{summary_pois is poisson regression summary}
\usage{
summary_pois(my_model_estimation, digit = 3)
}
\arguments{
\item{my_model_estimation}{poisson result}

\item{digit}{digit default 3}
}
\value{
tabular data for use in a paper
}
\description{
summary_pois is poisson regression summary
}
\examples{
\dontrun{
p <- read.csv("https://stats.idre.ucla.edu/stat/data/poisson_sim.csv")
p <- within(p, {
  prog <- factor(prog, levels=1:3,
                   labels=c("General", "Academic", "Vocational"))
                 id <- factor(id) })
m1 <- glm(num_awards ~ prog + math, family="poisson", data=p)
summary(m1)
Poisson_summary_function(m1)
}

}
