% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_report.R
\name{t_test_report}
\alias{t_test_report}
\title{Advanceed t test report function}
\usage{
t_test_report(data, iv = NULL, dv = NULL, type = "all")
}
\arguments{
\item{data}{data.frame}

\item{iv}{independent variable}

\item{dv}{dependent variable}

\item{type}{type option default is 'all'. and  'var.test', 'var.test.report', 't.test','t.test.report', 'boxplot', 'var.test.full', 't.test.full' output each result}
}
\description{
Advanceed t test report function
}
\examples{
# mtcars data
\dontrun{
t_test_report(mtcars, iv="vs",dv="mpg")
}
}
