% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizLCA.R
\name{vizLCA}
\alias{vizLCA}
\title{viz LCA Latent Class Analysis}
\usage{
vizLCA(
  model_lca,
  type = "res2",
  text_angle = 90,
  class_name = NULL,
  intercept = 10,
  Break = c(0, 20, 30, 50, 100),
  size_y_text = 10,
  sel = 1:20
)
}
\arguments{
\item{model_lca}{poLCA data}

\item{type}{res1, res2}

\item{text_angle}{text angle}

\item{class_name}{class name}

\item{intercept}{intercepts}

\item{Break}{c(0, 20,30, 50, 100)}

\item{size_y_text}{size =10}

\item{sel}{1:20 Select range if necessary}
}
\value{
LCA result, plot
}
\description{
viz LCA Latent Class Analysis
}
\examples{
\dontrun{

data(election)
library(glca)
data(gss08)
gss08na <- gss08 \%>\% drop_na()
gss08na
fm1 <- cbind(MORALG,CARESG,KNOWG,MORALB,CARESB,
             KNOWB,LEADB,DISHONB,INTELB)~ PARTY

poLCA(formula=fm1,data=elec,nclass=3,nrep=10,graphs = T)\%>\% vizLCA()

poLCA(cbind(DEFECT, HLTH,RAPE,POOR,SINGLE, NOMORE)~ SEX,
      data = gss08na, nclass = 3, graphs = TRUE ) \%>\% vizLCA()
poLCA1 <-  poLCA(cbind(DEFECT, HLTH,RAPE,POOR,SINGLE, NOMORE)~ SEX,
      data = gss08na, nclass = 3, graphs = TRUE ) \%>\% vizLCA()
poLCA1 \%>\% vizLCA()
poLCA1 \%>\% vizLCA(class_name = c("A","B","C"))
poLCA1 \%>\% vizLCA("df2") \%>\% print(n=Inf)
poLCA1 \%>\% vizLCA("g")
poLCA1 \%>\% vizLCA("pattern") \%>\%
  arrange( desc(observed)) \%>\% top_n(10)
poLCA1 \%>\% plot()
poLCA1 \%>\% vizLCA("g1")
poLCA1 \%>\% vizLCA("g2")
poLCA1 \%>\% vizLCA("g3")
poLCA1 \%>\% vizLCA("g4")
poLCA1 \%>\% vizLCA("class")
poLCA1 \%>\% vizLCA("pattern")
poLCA1 \%>\% vizLCA("pattern_data") \%>\% top_n(10)
poLCA1 \%>\% vizLCA("class") \%>\%
   dplyr::select(1,2,3,4,5,6,7,8,9)

model_n3
model_n3 \%>\% vizLCA()
model_n3 \%>\% vizLCA(class_name = c("낙태반대","조건부찬성","낙태찬성"))
model_n3 \%>\% vizLCA("res1",size_y_text = 12)
model_n3 \%>\% vizLCA("class")
model_n3 \%>\% vizLCA("df1")
model_n3 \%>\% vizLCA("df2")
model_n3 \%>\% vizLCA("df3")
model_n3 \%>\% vizLCA("df3")
model_n3 \%>\% vizLCA("g")
model_n3 \%>\% vizLCA("g1")
model_n3 \%>\% vizLCA("g2")
model_n3 \%>\% vizLCA("g3") #error nothing
model_n3 \%>\% vizLCA("g4") #error  nothing
model_n3 \%>\% vizLCA("pattern_data")
}
}
