% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareFit_diff.R
\name{CompareFit_diff}
\alias{CompareFit_diff}
\title{Comparison of Model Fit Indices of two or more structural equation models}
\usage{
CompareFit_diff(fit, ...)
}
\arguments{
\item{fit}{lavaan result and input fit1, fit2, ...}

\item{...}{lavaan result and input fit1, fit2, ...}
}
\description{
Comparison of Model Fit Indices of two or more structural equation models
}
