% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmeans_contrast.R
\name{emmeans_contrast}
\alias{emmeans_contrast}
\title{Two Way ANOVA emmeans_contrast}
\usage{
emmeans_contrast(
  aov_data,
  var = "",
  ...,
  auto = TRUE,
  type = "all",
  hjust = 0.1,
  vjust = -2
)
}
\arguments{
\item{aov_data}{aov data , example aov(len ~ supp*dose, data = ToothGrowth)}

\item{var}{contaast}

\item{...}{contrast}

\item{auto}{contrast auto}

\item{type}{overall result 'all','anova','emmeans','contrast', 'APA','confint','interaction'is plot}

\item{hjust}{text position horizontal}

\item{vjust}{text positin vertical}
}
\description{
Two Way ANOVA emmeans_contrast
}
\examples{

\dontrun{
## 2*3 ANOVA
aov(len ~ supp*dose, data = ToothGrowth) \%>\%
  emmeans_contrast()

ToothGrowth \%>\% str()
ToothGrowth$dose<- as.factor(ToothGrowth$dose)
aov(len ~ factor(dose)*supp, data = ToothGrowth) \%>\% tidy()

aov(len ~ dose*supp, data = ToothGrowth) \%>\%
  emmeans(specs = c("dose", "supp")) \%>\% data.frame()


aov(len ~ dose*supp, data = ToothGrowth) \%>\%
  emmeans_contrast(type = "emmeans")
aov(len ~ supp*dose, data = ToothGrowth) \%>\%
  emmeans_contrast()

}


}
