% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Freq_table.R
\name{Freq_table}
\alias{Freq_table}
\title{Freq_table  Frequency output function------------}
\usage{
Freq_table(
  data,
  format = "markdown",
  title = "Frequency output function",
  sort = FALSE,
  prop = TRUE,
  arrange = ""
)
}
\arguments{
\item{data}{data.frame}

\item{format}{markdown}

\item{title}{Frequency output function}

\item{sort}{data sort}

\item{prop}{add proportion}

\item{arrange}{desc}
}
\description{
Freq_table  Frequency output function------------
}
