% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTMT.R
\name{HTMT}
\alias{HTMT}
\title{The heterotrait-monotrait ratio of correlations (HTMT).
Title}
\usage{
HTMT(
  model = NA,
  dataset = NA,
  format = "markdown",
  digits = 3,
  sample.cov = NULL,
  missing = "listwise",
  ordered = NULL,
  absolute = TRUE,
  htmt2 = TRUE
)
}
\arguments{
\item{model}{lavaan model syntax}

\item{dataset}{data.frame dataset}

\item{format}{"markdown' and 'html' and %>% kableExtra::kable_classic()}

\item{digits}{round default 3 digits}

\item{sample.cov}{cov matrix}

\item{missing}{fill listwise}

\item{ordered}{default NULL}

\item{absolute}{default NULL}

\item{htmt2}{default NULL}
}
\description{
The heterotrait-monotrait ratio of correlations (HTMT).
Title
}
\examples{
\dontrun{


library(lavaan)
model1 <- '
  # measurement model
    ind60 =~ x1 + x2 + x3
    dem60 =~ y1 + y2 + y3 + y4
    dem65 =~ y5 + y6 + y7 + y8
  # regressions
    dem60 ~ ind60
    dem65 ~ ind60 + dem60
  # residual correlations
    y1 ~~ y5
    y2 ~~ y4 + y6
    y3 ~~ y7
    y4 ~~ y8
    y6 ~~ y8
'

fit1 <- sem(model1, data = PoliticalDemocracy)
# summary(fit2, standardized = TRUE)
HTMT(model = model1, dataset = PoliticalDemocracy)
}

}
