% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LavaanPlot.R
\name{LavaanPlot}
\alias{LavaanPlot}
\title{Lavaan plot draw}
\usage{
LavaanPlot(
  x,
  stand = TRUE,
  rotation = "LR",
  coefs = TRUE,
  col = "gray60",
  covs = TRUE
)
}
\arguments{
\item{x}{lavaan result}

\item{stand}{standardzied TRUE}

\item{rotation}{rotation 'LR','RL', 'TB','BT'}

\item{coefs}{TRUE}

\item{col}{default 'gray60'}

\item{covs}{TRUE}
}
\value{
plot
}
\description{
Lavaan plot draw
}
