% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysummaryBy.R
\name{mysummaryBy}
\alias{mysummaryBy}
\title{statistics by group}
\usage{
mysummaryBy(formula, data, add_var = NULL, stat = FALSE)
}
\arguments{
\item{formula}{general formula}

\item{data}{data.frame}

\item{add_var}{one sample name, Used to obtain statistics for one variable}

\item{stat}{TRUE t.test and aov(), thes stat ="t.test", or 'aov'}
}
\value{
Mea, SD, N, min, max
}
\description{
statistics by group
}
\examples{
\dontrun{
## Used to obtain statistics for one variable
mysummaryBy(mpg ~ 1, data = mtcars)
mysummaryBy(mpg ~ 1, data = mtcars, "mpg")
mysummaryBy(mpg ~ 1, data = mtcars, add_var = "mpg")

##The group variable is 2 or greater
mysummaryBy(mpg ~ vs, data = mtcars)
mysummaryBy(mpg ~ vs+am, data = mtcars)
mysummaryBy(mpg ~ vs+am+cyl, data = mtcars)

##statistic data output
mysummaryBy(mpg ~ 1, data = mtcars, "mpg", stat="t.test")
mysummaryBy(mpg ~ vs, data = mtcars, stat="t.test")
mysummaryBy(mpg ~ vs+am, data = mtcars, stat="aov")
mysummaryBy(mpg ~ vs*cyl, data = mtcars, stat="aov")

}

}
